requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.9.2
daphne==2.5.0
distlib==0.3.0
distro==1.4.0
Django==3.0.7
heroku==0.1.4
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==1.5
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
virtualbox==2.0.0
webencodings==0.5.1
whitenoise==5.1.0
zope.interface==5.1.0